example refactor and fixes
