+++ examples/proxy-forward-http/client/main.go
@@ -1,0 +1,20 @@
+package main
+
+import "github.com/zubairhamed/canopus"
+
+func main() {
+	conn, err := canopus.Dial("localhost:5683")
+	if err != nil {
+		panic(err.Error())
+	}
+
+	req := canopus.NewRequest(canopus.MessageConfirmable, canopus.Get, canopus.GenerateMessageID())
+	req.SetProxyURI("https://httpbin.org/get")
+
+	canopus.PrintMessage(req.GetMessage())
+	resp, err := conn.Send(req)
+	if err != nil {
+		println("err", err)
+	}
+	canopus.PrintMessage(resp.GetMessage())
+}
