+++ examples/proxy-forward-http/server/main.go
@@ -1,0 +1,12 @@
+package main
+
+import "github.com/zubairhamed/canopus"
+
+func main() {
+	server := canopus.NewServer()
+	server.ProxyOverHttp(true)
+
+	server.ListenAndServe(":5683")
+	<-make(chan struct{})
+
+}
