+++ examples/simple/client/main.go
